# web-app/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.30.0",
    "title": "StudiesPlanner 1 API",
    "description": "Utilize esta documentação para aulixio ao uso da API do sistema!"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/subject/{id}/exams": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as provas da disciplina",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "exams": {
                  "type": "array",
                  "example":
                  [{
                    "name": "Nome de identificação da prova",
                    "subject": {
                      "id": "IDDISCIPLINA",
                      "author": "IDUSUARIO"
                    },
                    "theme": "Tema da prova",
                    "date": "2020-07-28",
                    "create": "2020-06-10"
                  }]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar o id da disciplina na requisição para buscar as provas!"
                }
              }
            }
          }
        }
      }
    },
...
    "/exam/": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as provas",
        "description": "Retorna todas as provas de todas as disciplinas cadastradas para o usuário",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "exams": {
                  "type": "array",
                  "example":
                  [{
                    "name": "Nome de identificação da prova",
                    "subject": {
                      "id": "IDDISCIPLINA",
                      "author": "IDUSUARIO"
                    },
                    "theme": "Tema da prova",
                    "date": "2020-07-28",
                    "create": "2020-06-10"
                  }]
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "Criar prova",
        "description": "Criar uma prova para uma disciplina em específico",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": true,
            "example": "Prova Um",
            "description": "Nome de identificação da prova"
          },
          {
            "in": "body",
            "name": "subject",
            "required": true,
            "example": "abcd1234",
            "description": "Id da disciplina em que a prova pertence"
          },
          {
            "in": "body",
            "name": "date",
            "required": true,
            "example": "2020-08-22",
            "description": "Data de realização da prova"
          },
          {
            "in": "body",
            "name": "theme",
            "required": false,
            "example": "Memória principal e barramentos",
            "description": "Conteúdo que vai ser abordado na prova"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Prova criada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar uma prova para ser salva!"
                }
              }
            }
          }
        }
      }
    },
    "/exam/{id}": {
      "get": {
        "tags": [],
        "summary": "Retorna uma prova em específico",
        "description": "Retorna a prova correspondente ao id passado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Prova Um"
                },
                "subject": {
                  "type": "object",
                  "example": {
                    "id": "IDDISCIPLINA",
                    "author": "IDUSUARIO"
                  }
                },
                "theme": {
                  "type": "string",
                  "example": "Memória principal e barramentos"
                },
                "date": {
                  "type": "date",
                  "example": "2020-08-22"
                },
                "create": {
                  "type": "date",
                  "example": "2020-08-26"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "summary": "Atualizar prova",
        "description": "Permite atualizar nome, data de realização e conteúdo abordado da prova",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": false,
            "example": "Prova Um",
            "description": "Nome de identificação da prova"
          },
          {
            "in": "body",
            "name": "date",
            "required": false,
            "example": "2020-08-22",
            "description": "Data de realização da prova"
          },
          {
            "in": "body",
            "name": "theme",
            "required": false,
            "example": "Memória principal e barramentos",
            "description": "Conteúdo que vai ser abordado na prova"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Prova atualizada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Não é possível atualizar a disciplina da prova!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "summary": "Deletar prova",
        "description": "Deleta uma prova específica do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Prova deletada!"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    },
...
    "/lesson/": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as aulas",
        "description": "Retorna todas as aulas de todas as disciplinas cadastradas para o usuário",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "lessons": {
                  "type": "array",
                  "example":
                  [{
                    "subject": {
                      "id": "IDDISCIPLINA",
                      "author": "IDUSUARIO"
                    },
                    "duration": 90,
                    "realizationDays": [{
                      "day": "mon",
                      "start": "18:50"
                    }],
                    "quantity": 64,
                    "create": "2020-07-15"
                  }]
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "Criar aula",
        "description": "Cria uma aula para uma disciplina em específico",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "subject",
            "required": true,
            "example": "abcd12345",
            "description": "Id da disciplina em que a aula pertence"
          },
          {
            "in": "body",
            "name": "duration",
            "required": true,
            "example": 90,
            "description": "Tempo de duração da aula (em minutos)"
          },
          {
            "in": "body",
            "name": "realizationDays",
            "required": true,
            "example": [{
              "day": "mon",
              "start": "18:50"
            }],
            "description": "Um array de objetos contendo os dias de realização da aula. O objeto contem o atributo day, que representa qual dia da semana a aula vai ser realizada ['mon', 'tue', 'wed', 'thu', 'fri', 'sat'], e o outro atributo é start que determina o horário de início da aula"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Aula criada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar uma aula para ser salva!"
                }
              }
            }
          }
        }
      }
    },
    "/lesson/{id}": {
      "get": {
        "tags": [],
        "summary": "Retorna uma aula em específico",
        "description": "Retorna a aula correspondente ao id passado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "subject": {
                  "type": "object",
                  "example": {
                    "id": "IDDISCIPLINA",
                    "author": "IDUSUARIO"
                  }
                },
                "duration": {
                  "type": "number",
                  "example": 90
                },
                "realizationDays": {
                  "type": "object",
                  "example": [{
                  "day": "mon",
                  "start": "18:50"
                }]},
                "quantity": {
                  "type": "number",
                  "example": 64
                },
                "create": {
                  "type": "date",
                  "example": "2020-07-10"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "summary": "Atualizar aula",
        "description": "Permite atualizar os dias de realização da aula e a duração",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "duration",
            "required": false,
            "example": 90,
            "description": "Tempo de duração da aula (em minutos)"
          },
          {
            "in": "body",
            "name": "realizationDays",
            "required": false,
            "example": [{
              "day": "mon",
              "start": "18:50"
            }],
            "description": "Um array de objetos contendo os dias de realização da aula. O objeto contem o atributo day, que representa qual dia da semana a aula vai ser realizada ['mon', 'tue', 'wed', 'thu', 'fri', 'sat'], e o outro atributo é start que determina o horário de início da aula"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Aula atualizada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Não é possível atualizar a disciplina da aula!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "summary": "Deletar aula",
        "description": "Deleta uma aula específica do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Aula deletada!"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "parameters": {
    "x-key-api": {
      "name": "x-key-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    },
    "x-token-api": {
      "name": "x-token-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    }
  }
}